api/powerplant/requirements.txt
# Flask框架
Flask==2.3.3

# 数据库相关
SQLAlchemy==2.0.23
PyMySQL==1.1.0

# HTTP请求
requests==2.31.0

# 数据处理
pandas==2.1.3
numpy==1.24.3

# 配置文件处理
PyYAML==6.0.1

# 文档生成
python-docx==0.8.11
reportlab==4.0.7

# 环境变量
python-dotenv==1.0.0
